refactor(NavBar): convert NavBar to hooks and rename NavBarContainer file
